Divorce tests from class and add skeleton app
