+++ BusinessLogic/Models/Player.cs
@@ -3,6 +3,7 @@
     public class Player
     {
         public Guid Id { get; set; }
+
         public string Name { get; set; }
     }
 }
