+++ EntityModel/Role.cs
@@ -3,7 +3,9 @@
     public class Role
     {
         public int Id { get; set; }
+
         public string DisplayName { get; set; } = string.Empty;
+
         public Team Team { get; set; }
     }
 }
