+++ EntityModel/Team.cs
@@ -3,6 +3,7 @@
     public enum Team
     {
         Resistance = 0,
+
         Spies = 1
     }
 }
