+++ ResistanceRoleGenerator.sln
@@ -3,9 +3,15 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32602.215
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BusinessLogic", "BusinessLogic\BusinessLogic.csproj", "{0E100736-CDD4-4A0A-BBA2-037827DFFB7A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogic", "BusinessLogic\BusinessLogic.csproj", "{0E100736-CDD4-4A0A-BBA2-037827DFFB7A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EntityModel", "EntityModel\EntityModel.csproj", "{AE2B3155-B144-4777-A999-E06424B48B8E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Resistance", "Resistance\Resistance.csproj", "{B122FF8C-247F-48D4-9DAF-06B3E60446D3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogicTests", "BusinessLogicTests\BusinessLogicTests.csproj", "{A8846BDF-0ED3-4C31-9211-6A4F6D9FFABC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{E4E1C578-E487-4BE9-97E2-598D277B6570}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,11 +27,22 @@ Global
 		{AE2B3155-B144-4777-A999-E06424B48B8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AE2B3155-B144-4777-A999-E06424B48B8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AE2B3155-B144-4777-A999-E06424B48B8E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B122FF8C-247F-48D4-9DAF-06B3E60446D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B122FF8C-247F-48D4-9DAF-06B3E60446D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B122FF8C-247F-48D4-9DAF-06B3E60446D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B122FF8C-247F-48D4-9DAF-06B3E60446D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8846BDF-0ED3-4C31-9211-6A4F6D9FFABC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8846BDF-0ED3-4C31-9211-6A4F6D9FFABC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8846BDF-0ED3-4C31-9211-6A4F6D9FFABC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8846BDF-0ED3-4C31-9211-6A4F6D9FFABC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E0A0469B-7718-4100-8703-505FE699BE6B}
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{A8846BDF-0ED3-4C31-9211-6A4F6D9FFABC} = {E4E1C578-E487-4BE9-97E2-598D277B6570}
 	EndGlobalSection
 EndGlobal
